Move linear attention in layers/mamba
